fix: cr opt
