+++ packages/core/src/physics/joint/FixedJoint.ts
@@ -9,23 +9,6 @@ import { IFixedJoint } from "@oasis-engine/design";
  */
 export class FixedJoint extends Joint {
   /**
-   * The connected anchor position.
-   * @remarks If connectedCollider is set, this anchor is relative offset.
-   * Or the anchor is world anchor position.
-   */
-  get connectedAnchor(): Vector3 {
-    return this._connectedCollider.localPosition;
-  }
-
-  set connectedAnchor(value: Vector3) {
-    const connectedAnchor = this._connectedCollider.localPosition;
-    if (value !== connectedAnchor) {
-      connectedAnchor.copyFrom(value);
-    }
-    (<IFixedJoint>this._nativeJoint).setConnectedAnchor(value);
-  }
-
-  /**
    * @override
    * @internal
    */
