+++ packages/core/src/physics/joint/SpringJoint.ts
@@ -30,23 +30,6 @@ export class SpringJoint extends Joint {
   }
 
   /**
-   * The connected anchor position.
-   * @remarks If connectedCollider is set, this anchor is relative offset.
-   * Or the anchor is world anchor position.
-   */
-  get connectedAnchor(): Vector3 {
-    return this._connectedCollider.localPosition;
-  }
-
-  set connectedAnchor(value: Vector3) {
-    const connectedAnchor = this._connectedCollider.localPosition;
-    if (value !== connectedAnchor) {
-      connectedAnchor.copyFrom(value);
-    }
-    (<ISpringJoint>this._nativeJoint).setConnectedAnchor(value);
-  }
-
-  /**
    * The minimum distance.
    */
   get minDistance(): number {
