+++ packages/design/src/physics/joints/IJoint.ts
@@ -11,6 +11,12 @@ export interface IJoint {
   setConnectedCollider(value: ICollider): void;
 
   /**
+   * The connected anchor position.
+   * @remarks If connectedCollider is set, this anchor is relative offset, or the anchor is world position.
+   */
+  setConnectedAnchor(value: Vector3): void;
+
+  /**
    *  The scale to apply to the inverse mass of collider 0 for resolving this constraint.
    */
   setConnectedMassScale(value: number): void;
